Add files via upload
JavaScript(often shortened to JS) is a lightweight, interpreted, object-oriented language with first-class functions, and is best known as the scripting language for Web pages, but it's used in many non-browser environments as well. JavaScript can function as both a procedural and an object oriented language. A calculator, as we know, will perform our basic operations viz. Addition, Subtraction, Multiplication, and division. For starting out, The Technologies used for designing the designing are done  with HTML and CSS.